requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
python-dotenv==1.0.0
gunicorn==20.1.0
psycopg2-binary
waitress==2.1.2 